+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.4.9] - 2026-04-07
+
+### Fixed
+- Handle Kraken execeptions
+- Kraken spot candles request
+
 ## [1.4.8] - 2026-03-24
 
 ### Changed